style(选择数据/数据库): 提取请求逻辑.
